pyproject.toml
[tool.poetry]
name = "books-crud-api"
version = "0.1.0"
description = ""
authors = ["Georgy Bachakashvili"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
django-rest-framework = "^0.1.0"
djangorestframework-csv = "^3.0.2"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.8"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
